+++ Scroller/EditableWrapper.js
@@ -4,9 +4,9 @@ import {is} from '@enact/core/keymap';
 import {usePublicClassNames} from '@enact/core/usePublicClassNames';
 import Spotlight, {getDirection} from '@enact/spotlight';
 import {getContainersForNode} from '@enact/spotlight/src/container';
-import {getTargetByDirectionFromElement} from '@enact/spotlight/src/target';
+import {getTargetByDirectionFromElement, getTargetByDirectionFromPosition} from '@enact/spotlight/src/target';
 import Accelerator from '@enact/spotlight/Accelerator';
-import {getLastPointerPosition, setPointerMode} from '@enact/spotlight/src/pointer';
+import {getLastPointerPosition, getPointerMode, setPointerMode} from '@enact/spotlight/src/pointer';
 import {Announce} from '@enact/ui/AnnounceDecorator';
 import Touchable from '@enact/ui/Touchable';
 import classNames from 'classnames';
@@ -113,8 +113,8 @@ const EditableWrapper = (props) => {
 		// Position for restoring focus after removing item
 		nextSpotlightRect: null,
 
-		// Last move direction
-		lastMoveDirection: null,
+		// Move direction from the starting position to the current position
+		moveDirection: null,
 
 		// Last mouse position
 		lastMouseClientX: null,
@@ -151,7 +151,7 @@ const EditableWrapper = (props) => {
 		mutableRef.current.focusedItem = null;
 		mutableRef.current.selectedItem = null;
 		mutableRef.current.selectedItemLabel = '';
-		mutableRef.current.lastMoveDirection = null;
+		mutableRef.current.moveDirection = null;
 		mutableRef.current.prevToIndex = null;
 		mutableRef.current.initialSelected = null;
 		wrapperRef.current.style.setProperty('--selected-item-offset', '0px');
@@ -161,7 +161,7 @@ const EditableWrapper = (props) => {
 
 	// Finalize the order
 	const finalizeOrders = useCallback(() => {
-		const {fromIndex, lastMoveDirection, prevToIndex, rearrangedItems, selectedItem} = mutableRef.current;
+		const {fromIndex, moveDirection, prevToIndex, rearrangedItems, selectedItem} = mutableRef.current;
 		let orders = Array.from({length: dataSize}, (_, i) => i + 1);
 
 		if (rearrangedItems.length > 0) {
@@ -171,17 +171,17 @@ const EditableWrapper = (props) => {
 			rearrangedItems.forEach((item) => {
 				const order = Number(item.style.order);
 				selectedOrder = order;
-				item.style.order = order - lastMoveDirection;
+				item.style.order = order - moveDirection;
 				item.classList.remove(componentCss.rearrangedTransform, componentCss.rearranged);
 
-				if (lastMoveDirection > 0) {
+				if (moveDirection > 0) {
 					changedOrder.push(order);
 				} else {
 					changedOrder.unshift(order);
 				}
 			});
 
-			if (lastMoveDirection > 0) {
+			if (moveDirection > 0) {
 				changedOrder.push(Number(selectedItem.style.order));
 			} else {
 				changedOrder.unshift(Number(selectedItem.style.order));
@@ -319,17 +319,17 @@ const EditableWrapper = (props) => {
 	}, []);
 
 	// Add rearranged items
-	const addRearrangedItems = useCallback(({moveDirection, toIndex}) => {
-		// Set the moveDirection to css variable
+	const addRearrangedItems = useCallback(({currentMoveDirection, toIndex}) => {
+		// Set the currentMoveDirection to css variable
 		const {rtl} = scrollContainerHandle.current;
-		wrapperRef.current.style.setProperty('--move-direction', moveDirection * (rtl ? -1 : 1));
+		wrapperRef.current.style.setProperty('--move-direction', currentMoveDirection * (rtl ? -1 : 1));
 
 		const {fromIndex, rearrangedItems, selectedItem} = mutableRef.current;
-		const getNextElement = (item) => moveDirection > 0 ? item.nextElementSibling : item.previousElementSibling;
+		const getNextElement = (item) => currentMoveDirection > 0 ? item.nextElementSibling : item.previousElementSibling;
 		let sibling = getNextElement(selectedItem);
 		let lastRearrangedItem;
-		let start = moveDirection > 0 ? toIndex : fromIndex;
-		let end = moveDirection > 0 ? fromIndex : toIndex;
+		let start = currentMoveDirection > 0 ? toIndex : fromIndex;
+		let end = currentMoveDirection > 0 ? fromIndex : toIndex;
 
 		while (start > end && sibling) {
 			sibling?.classList.add(componentCss.rearranged, componentCss.rearrangedTransform);
@@ -347,7 +347,7 @@ const EditableWrapper = (props) => {
 			readOutCurrentPosition(lastRearrangedItem.ariaLabel || lastRearrangedItem.textContent);
 		}
 
-		mutableRef.current.lastMoveDirection = moveDirection;
+		mutableRef.current.moveDirection = currentMoveDirection;
 
 	}, [readOutCurrentPosition, scrollContainerHandle]);
 
@@ -374,7 +374,7 @@ const EditableWrapper = (props) => {
 		if (selectedItem) {
 			// Bail out when index is out of scope
 			if (toIndex < mutableRef.current.hideIndex && toIndex >= 0) {
-				const {fromIndex, itemWidth, lastMoveDirection, prevToIndex, rearrangedItems} = mutableRef.current;
+				const {fromIndex, itemWidth, moveDirection, prevToIndex, rearrangedItems} = mutableRef.current;
 
 				// Set the selected item's offset to css variable
 				const offset = (toIndex - fromIndex) * itemWidth;
@@ -383,18 +383,19 @@ const EditableWrapper = (props) => {
 				// If the current toIndex is new,
 				if (toIndex !== prevToIndex) {
 					// Determine the direction of move from the latest from index
-					const moveDirection = Math.sign(toIndex - prevToIndex);
+					const currentMoveDirection = Math.sign(toIndex - prevToIndex);
 					// If the direction is changed and there are rearranged items, we remove them first.
-					if (lastMoveDirection && moveDirection !== lastMoveDirection && rearrangedItems.length > 0) {
-						const numToRemove = moveDirection > 0 ? toIndex - prevToIndex : prevToIndex - toIndex;
+					if (moveDirection && currentMoveDirection !== moveDirection && rearrangedItems.length > 0) {
+						const numToRemove = currentMoveDirection > 0 ? toIndex - prevToIndex : prevToIndex - toIndex;
+						const needToAddItems = numToRemove > rearrangedItems.length;
 						removeRearrangedItems(numToRemove);
 
 						// When there's jump, meaning, numToRemove is bigger than 0, we need to add an item
-						if (numToRemove > rearrangedItems.length) {
-							addRearrangedItems({moveDirection, toIndex});
+						if (needToAddItems) {
+							addRearrangedItems({currentMoveDirection, toIndex});
 						}
 					} else {
-						addRearrangedItems({moveDirection, toIndex});
+						addRearrangedItems({currentMoveDirection, toIndex});
 					}
 
 					mutableRef.current.prevToIndex = toIndex;
@@ -520,8 +521,8 @@ const EditableWrapper = (props) => {
 		// Coordinate calculation in RTL locales is not supported in chrome below 85
 		const scrollContentOffset = scrollContentRef.current.scrollLeft * (rtl ? -1 : 1) - centeredOffset;
 		const clientXFromContent = (rtl ? bodyWidth - x : x) + scrollContentOffset;
-		const moveDirection = (itemWidth * (prevToIndex + 0.5) < clientXFromContent) ? 1 : -1; // 1: To next index , -1: To prev index
-		const moveTolerance = itemWidth * tolerance * moveDirection;
+		const direction = (itemWidth * (prevToIndex + 0.5) < clientXFromContent) ? 1 : -1; // 1: To next index , -1: To prev index
+		const moveTolerance = itemWidth * tolerance * direction;
 
 		return Math.floor((clientXFromContent - moveTolerance) / itemWidth);
 	}, [scrollContainerHandle, scrollContentRef]);
@@ -558,30 +559,36 @@ const EditableWrapper = (props) => {
 		}
 	}, [finalizeEditing, finalizeOrders, scrollContainerHandle, scrollContentRef]);
 
+	const completeEditingByKeyDown = useCallback((target) => {
+		const {selectedItem, selectedItemLabel} = mutableRef.current;
+		const orders = finalizeOrders();
+
+		finalizeEditing(orders);
+		if (selectItemBy === 'press') {
+			focusItem(target);
+		}
+		selectedItem.children[1].ariaLabel = '';
+		setTimeout(() => {
+			announceRef.current.announce(
+				selectedItemLabel + $L('Movement completed'),
+				true
+			);
+			setTimeout(() => {
+				selectedItem.children[1].ariaLabel = `${selectedItem.ariaLabel} ${$L('Press the OK button to move or press the up button to select other options.')}`;
+			}, completeAnnounceDelay);
+		}, completeAnnounceDelay);
+	}, [finalizeEditing, finalizeOrders, focusItem, selectItemBy]);
+
 	const handleKeyDownCapture = useCallback((ev) => {
 		const {keyCode, repeat, target} = ev;
-		const {focusedItem, selectedItem, selectedItemLabel} = mutableRef.current;
+		const {focusedItem, selectedItem} = mutableRef.current;
 		const targetItemNode = findItemNode(target);
 
 		if (is('enter', keyCode) && target.getAttribute('role') !== 'button') {
 			if (!repeat) {
 				if (selectedItem) {
-					const orders = finalizeOrders();
-					finalizeEditing(orders);
-					if (selectItemBy === 'press') {
-						focusItem(target);
-					}
+					completeEditingByKeyDown(target);
 					mutableRef.current.needToPreventEvent = true;
-					selectedItem.children[1].ariaLabel = '';
-					setTimeout(() => {
-						announceRef.current.announce(
-							selectedItemLabel + $L('Movement completed'),
-							true
-						);
-						setTimeout(() => {
-							selectedItem.children[1].ariaLabel = `${selectedItem.ariaLabel} ${$L('Press the OK button to move or press the up button to select other options.')}`;
-						}, completeAnnounceDelay);
-					}, completeAnnounceDelay);
 				} else if (selectItemBy === 'press') {
 					startEditing(targetItemNode);
 					mutableRef.current.needToPreventEvent = true;
@@ -591,6 +598,9 @@ const EditableWrapper = (props) => {
 					startEditing(targetItemNode);
 				}, holdDuration - 300);
 			}
+		} else if (is('down', keyCode) && target.getAttribute('role') !== 'button' && !repeat && selectedItem) {
+			completeEditingByKeyDown(target);
+			mutableRef.current.needToPreventEvent = true;
 		} else if (is('left', keyCode) || is('right', keyCode)) {
 			if (selectedItem) {
 				if (mutableRef.current.lastKeyEventTargetElement?.getAttribute('role') !== 'button') {
@@ -619,6 +629,7 @@ const EditableWrapper = (props) => {
 
 				// Check if focus leaves scroll container.
 				if (nextTarget && !getContainersForNode(nextTarget).includes(mutableRef.current.spotlightId)) {
+					setPointerMode(false);
 					Spotlight.move(getDirection(keyCode));
 
 					const orders = finalizeOrders();
@@ -629,30 +640,15 @@ const EditableWrapper = (props) => {
 				}
 			}
 		}
-	}, [finalizeEditing, finalizeOrders, findItemNode, focusItem, moveItemsByKeyDown, reset, selectItemBy, startEditing]);
+	}, [finalizeEditing, finalizeOrders, findItemNode, moveItemsByKeyDown, reset, selectItemBy, startEditing, completeEditingByKeyDown]);
 
 	const handleKeyUpCapture = useCallback((ev) => {
 		const {keyCode, target} = ev;
-		const {selectedItem, selectedItemLabel} = mutableRef.current;
+		const {selectedItem} = mutableRef.current;
 
 		if (is('cancel', keyCode)) {
 			if (selectedItem) {
-				const orders = finalizeOrders();
-				finalizeEditing(orders);
-				if (selectItemBy === 'press') {
-					focusItem(target);
-				}
-				selectedItem.children[1].ariaLabel = '';
-				setTimeout(() => {
-					announceRef?.current?.announce(
-						selectedItemLabel + $L('Movement completed'),
-						true
-					);
-					setTimeout(() => {
-						selectedItem.children[1].ariaLabel = `${selectedItem.ariaLabel} ${$L('Press the OK button to move or press the up button to select other options.')}`;
-					}, completeAnnounceDelay);
-				}, completeAnnounceDelay);
-
+				completeEditingByKeyDown(target);
 				ev.stopPropagation(); // To prevent onCancel by CancelDecorator
 			}
 		} else {
@@ -668,7 +664,37 @@ const EditableWrapper = (props) => {
 			ev.preventDefault();
 			mutableRef.current.needToPreventEvent = false;
 		}
-	}, [finalizeEditing, finalizeOrders, focusItem, selectItemBy]);
+	}, [completeEditingByKeyDown]);
+
+	const handleGlobalKeyDownCapture = useCallback((ev) => {
+		if (getPointerMode() && !scrollContainerRef.current.contains(Spotlight.getCurrent()) && mutableRef.current.selectedItem) {
+			const {keyCode} = ev;
+			const position = getLastPointerPosition();
+			const direction = getDirection(keyCode);
+			if (direction) {
+				const nextTarget = getTargetByDirectionFromPosition(direction, position, mutableRef.current.spotlightId);
+
+				if (!scrollContainerRef.current.contains(nextTarget)) {
+					const orders = finalizeOrders();
+					finalizeEditing(orders);
+				}
+			}
+		}
+	}, [finalizeEditing, finalizeOrders, scrollContainerRef]);
+
+	useLayoutEffect(() => {
+		const available = typeof document === 'object';
+
+		if (available) {
+			document.addEventListener('keydown', handleGlobalKeyDownCapture, {capture: true});
+		}
+
+		return () => {
+			if (available) {
+				document.removeEventListener('keydown', handleGlobalKeyDownCapture, {capture: true});
+			}
+		};
+	}, [handleGlobalKeyDownCapture]);
 
 	useEffect(() => {
 		if (mutableRef.current.nextSpotlightRect !== null) {